--- a/.changeset/good-ghosts-grin.md
+++ b/.changeset/good-ghosts-grin.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/live-common": patch
+---
+
+Implement useThrottledPortfolio hook to allow throttling the computation of portfolio graph data