--- a/apps/ledger-live-mobile/src/analytics/index.ts
+++ b/apps/ledger-live-mobile/src/analytics/index.ts
@@ -1,7 +1,5 @@
-// eslint-disable-next-line import/no-cycle
 import Track from "./Track";
-// eslint-disable-next-line import/no-cycle
 import TrackScreen from "./TrackScreen";
-// eslint-disable-next-line import/no-cycle
+
 export * from "./segment";
 export { Track, TrackScreen };