--- a/apps/ledger-live-mobile/src/components/AssetCentricGraphCard.tsx
+++ b/apps/ledger-live-mobile/src/components/AssetCentricGraphCard.tsx
@@ -14,7 +14,7 @@ import Delta from "./Delta";
 import CurrencyUnitValue from "./CurrencyUnitValue";
 import getWindowDimensions from "../logic/getWindowDimensions";
 import Graph from "./Graph";
-import TransactionsPendingConfirmationWarning from "./TransactionsPendingConfirmationWarning";
+import { TransactionsPendingConfirmationWarningAllAccounts } from "./TransactionsPendingConfirmationWarning";
 import ParentCurrencyIcon from "./ParentCurrencyIcon";
 import FormatDate from "./FormatDate";
 import { ensureContrast } from "../colors";
@@ -182,7 +182,7 @@ function AssetCentricGraphCard({
                     </Text>
                   </Flex>
                 )}
-                <TransactionsPendingConfirmationWarning />
+                <TransactionsPendingConfirmationWarningAllAccounts />
               </Flex>
               <Flex flexDirection={"row"}>
                 {!balanceHistory ? (