--- a/apps/ledger-live-mobile/src/components/FormatDate.tsx
+++ b/apps/ledger-live-mobile/src/components/FormatDate.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useSelector } from "react-redux";
+import { createSelector } from "reselect";
 import compareDate from "../logic/compareDate";
 import { localeSelector } from "../reducers/settings";
 
@@ -7,27 +8,37 @@ type Props = {
   date: Date | null | undefined;
   withHoursMinutes?: boolean;
 };
-const defaultOptions: Intl.DateTimeFormatOptions = {
-  year: "numeric",
-  month: "numeric",
-  day: "numeric",
-};
-const hoursAndMinutesOptions: Intl.DateTimeFormatOptions = {
-  hour: "2-digit",
-  minute: "2-digit",
-};
+
+const defaultOptionsSelector = createSelector(
+  localeSelector,
+  locale =>
+    new Intl.DateTimeFormat(locale, {
+      year: "numeric",
+      month: "numeric",
+      day: "numeric",
+    }),
+);
+
+const hoursAndMinutesOptionsSelector = createSelector(
+  localeSelector,
+  locale =>
+    new Intl.DateTimeFormat(locale, {
+      hour: "2-digit",
+      minute: "2-digit",
+    }),
+);
 
 function FormatDate({
   date,
   withHoursMinutes = false,
 }: Props): JSX.Element | null {
-  const locale = useSelector(localeSelector);
+  const defaultOptions = useSelector(defaultOptionsSelector);
+  const hoursAndMinutesOptions = useSelector(hoursAndMinutesOptionsSelector);
   const jsx =
     date && date.getTime()
-      ? new Intl.DateTimeFormat(locale, {
-          ...defaultOptions,
-          ...(withHoursMinutes ? hoursAndMinutesOptions : {}),
-        }).format(date)
+      ? withHoursMinutes
+        ? defaultOptions.format(date)
+        : hoursAndMinutesOptions.format(date)
       : null;
   return <>{jsx}</>;
 }