--- a/apps/ledger-live-mobile/src/components/OperationIcon.tsx
+++ b/apps/ledger-live-mobile/src/components/OperationIcon.tsx
@@ -89,4 +89,4 @@ export default connect((state: State, props: OwnProps) => {
     type,
     confirmed: isConfirmed,
   };
-})(OperationIcon);
+})(React.memo(OperationIcon));