--- a/apps/ledger-live-mobile/src/components/OperationRow.tsx
+++ b/apps/ledger-live-mobile/src/components/OperationRow.tsx
@@ -13,6 +13,7 @@ import {
 import { Account, Operation, AccountLike } from "@ledgerhq/types-live";
 import { Box, Flex, InfiniteLoader, Text } from "@ledgerhq/native-ui";
 import debounce from "lodash/debounce";
+import { isEqual } from "lodash";
 import CurrencyUnitValue from "./CurrencyUnitValue";
 import CounterValue from "./CounterValue";
 import OperationIcon from "./OperationIcon";
@@ -81,7 +82,7 @@ const placeholderProps = {
   containerHeight: 20,
 };
 
-export default function OperationRow({
+function OperationRow({
   account,
   parentAccount,
   operation,
@@ -252,3 +253,9 @@ export default function OperationRow({
     </ContainerTouchable>
   );
 }
+
+/**
+ * you might be surprised but isEqual on these objects is actually very fast
+ * (are we keeping the same object refs at a deep level? no lo se señor)
+ * */
+export default React.memo(OperationRow, isEqual);