--- a/apps/ledger-live-mobile/src/components/OperationRowDate.tsx
+++ b/apps/ledger-live-mobile/src/components/OperationRowDate.tsx
@@ -1,19 +1,26 @@
 import React, { useMemo } from "react";
 import { useSelector } from "react-redux";
+import { createSelector } from "reselect";
 import { localeSelector } from "../reducers/settings";
 
 type Props = {
   date: Date;
 };
+
+const localeDateTimeFormatSelector = createSelector(
+  localeSelector,
+  locale =>
+    new Intl.DateTimeFormat(locale, {
+      hour: "2-digit",
+      minute: "2-digit",
+    }),
+);
+
 export default function OperationRowDate({ date }: Props) {
-  const locale = useSelector(localeSelector);
+  const dateTimeFormat = useSelector(localeDateTimeFormatSelector);
   const localeTimeString = useMemo(
-    () =>
-      date.toLocaleTimeString(locale, {
-        hour: "2-digit",
-        minute: "2-digit",
-      }),
-    [date, locale],
+    () => dateTimeFormat.format(date),
+    [date, dateTimeFormat],
   );
   return <>{localeTimeString}</>;
 }