--- a/apps/ledger-live-mobile/src/hooks/portfolio.ts
+++ b/apps/ledger-live-mobile/src/hooks/portfolio.ts
@@ -6,8 +6,8 @@ import type {
 } from "@ledgerhq/types-cryptoassets";
 import {
   useBalanceHistoryWithCountervalue as useBalanceHistoryWithCountervalueCommon,
-  usePortfolio as usePortfolioCommon,
   useCurrencyPortfolio as useCurrencyPortfolioCommon,
+  usePortfolioThrottled,
 } from "@ledgerhq/live-common/portfolio/v2/react";
 import { GetPortfolioOptionsType } from "@ledgerhq/live-common/portfolio/v2/index";
 import {
@@ -31,15 +31,26 @@ export function useBalanceHistoryWithCountervalue({
   });
 }
 
-export function usePortfolio(
-  accounts?: AccountLike[],
+export function usePortfolioAllAccounts(options?: GetPortfolioOptionsType) {
+  const to = useSelector(counterValueCurrencySelector);
+  const accounts = useSelector(accountsSelector);
+  const range = useSelector(selectedTimeRangeSelector);
+  return usePortfolioThrottled({
+    accounts,
+    range,
+    to,
+    options,
+  });
+}
+
+export function usePortfolioForAccounts(
+  accounts: AccountLike[],
   options?: GetPortfolioOptionsType,
 ) {
   const to = useSelector(counterValueCurrencySelector);
-  const accountsSelected = useSelector(accountsSelector);
   const range = useSelector(selectedTimeRangeSelector);
-  return usePortfolioCommon({
-    accounts: accounts || accountsSelected,
+  return usePortfolioThrottled({
+    accounts,
     range,
     to,
     options,