--- a/apps/ledger-live-mobile/src/screens/Account/ReadOnly/ReadOnlyAccount.tsx
+++ b/apps/ledger-live-mobile/src/screens/Account/ReadOnly/ReadOnlyAccount.tsx
@@ -172,4 +172,4 @@ function ReadOnlyAccount({ route }: Props) {
   );
 }
 
-export default withDiscreetMode(ReadOnlyAccount);
+export default React.memo(withDiscreetMode(ReadOnlyAccount));