--- a/apps/ledger-live-mobile/src/screens/Account/index.tsx
+++ b/apps/ledger-live-mobile/src/screens/Account/index.tsx
@@ -226,4 +226,4 @@ const AccountScreenInner = ({
   );
 };
 
-export default withDiscreetMode(AccountScreen);
+export default React.memo(withDiscreetMode(AccountScreen));