--- a/apps/ledger-live-mobile/src/screens/Analytics/Operations.tsx
+++ b/apps/ledger-live-mobile/src/screens/Analytics/Operations.tsx
@@ -113,7 +113,7 @@ export function Operations({ navigation, route }: Props) {
   }: {
     section: SectionListData<Operation, DailyOperationsSection>;
   }) {
-    return <SectionHeader section={section} />;
+    return <SectionHeader day={section.day} />;
   }
 
   const onTransactionButtonPress = useCallback(() => {