… .changeset/purple-files-enjoy.md
… apps/ledger-live-mobile/src/components/Carousel/Slide.tsx
… apps/ledger-live-mobile/src/components/Carousel/shared.tsx
… apps/ledger-live-mobile/src/components/FormatDay.ts
… apps/ledger-live-mobile/src/screens/Analytics/RingChart.tsx
… apps/ledger-live-mobile/src/screens/CustomImage/NFTGallerySelector.tsx
… apps/ledger-live-mobile/src/screens/Lending/History/index.tsx
… apps/ledger-live-mobile/src/screens/Nft/NftCollection/index.tsx
… apps/ledger-live-mobile/src/screens/Nft/WalletNftGallery/index.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/Assets.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/Header.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/PortfolioAssets.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/PortfolioGraphCard.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/PortfolioHistory.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/PortfolioOperationsHistorySection.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/ReadOnly/index.tsx
… apps/ledger-live-mobile/src/screens/Portfolio/index.tsx
… apps/ledger-live-mobile/src/screens/Settings/Debug/Features/index.tsx
… apps/ledger-live-mobile/src/screens/Settings/Debug/Generators/ImportBridgeStreamData.tsx
… apps/ledger-live-mobile/src/screens/Settings/Debug/Generators/ToggleServiceStatus.tsx
… apps/ledger-live-mobile/src/screens/Settings/Debug/Generators/index.tsx
… apps/ledger-live-mobile/src/screens/Settings/index.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/AssetDynamicContent.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/AssetGraph.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/AssetMarketSection.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/AssetRow.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/Header.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/ReadOnly/index.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricAsset/index.tsx
… apps/ledger-live-mobile/src/screens/WalletCentricSections/Allocations.tsx