<!--
Thank you for your contribution! 👍
Please make sure to read CONTRIBUTING.md if you have not already.
Disclaimer: Pull Requests that do not comply with the rules will be arbitrarily closed.
-->

### 📝 Description

In the context of an audit of the performance in LLM, I did some profiling (mainly React profiling with the dev tools).

- This PR addresses one of the main issues observed: **high frequency of store updates causing re-renders of entire screens**.
  - If a user has N accounts, when the app syncs these N accounts, it results in N successive updates of the `store.getState().accounts.active` part of the store, they happen at a high frequency (in the order of 1 per second). **This frequency of updates was not directly addressed here, but ...**
  - How it impacts the UI performance:
  In many components of the app we have `useSelector(accountsSelector)` that points directly to that part of the store, so each of these updates in the store triggers a re-render of all these components. We are talking about **entire screens that get fully re-rendered very frequently** with a cost of ~200ms (in debug) per render, that's a lot of frames dropped.
  - How it is addressed:
  **In a lot of cases these components only consume data that is _derived from_ `accounts.active`**, not the entire data. For instance in a lot of cases we just count the number of accounts or whether some accounts have a positive balance.
  **The fix is then to create selectors that directly derive the data needed**, thus restricting re-renders to updates that cause this derived data to change.

- Another directly related issue is the **expensive computation of graph data**.
  - The components rendering the graph actually need the full data coming from `accounts.active` (and other frequently sources of data such as the countervalues state), so there, optimising the selectors is not be the right strategy (also it's more difficult because the data of countervalues comes from a context, not from the store).
  - It impacts the UI performance because each of these computation happens synchronously, can take ~50ms (in debug) and thus block the JS thread for several frames, making the app less responsive.
  - The solution here is to throttle the inputs of that computation so that a computation happens at most every 5000ms instead of at every single render.  For that purpose several `useThrottle*` hooks were implemented in ledger-live-common and allow throttling a value, an array of values, or the execution of a function along with its dependencies.

- Another important thing that was implemented here is to **break down big components into smaller components** so that re-renders only happen in parts that actually needs this re-render. The general idea is to eliminate as much state as possible from big components and pull it down when possible (when that state isn't shared between the smaller parts). This is especially possible on the main Portfolio screen and the WalletCentricAsset screen.

#### Scope of changes in UI

- Portfolio main screen (including ReadOnly)
- Asset main screen (including ReadOnly)
- NFT gallery
- Custom image NFT gallery selector
- Main list of operations
- Settings main screen (Accounts section)
- Settings debug: Toggle Service status incident section

#### Important points to fully understand the changes in this PR:

- [How `useSelector` works and when does it trigger re-renders](https://react-redux.js.org/api/hooks#useselector): 
> <img width="834" alt="Screenshot 2023-02-03 at 10 28 04" src="https://user-images.githubusercontent.com/91890529/216563658-cf70ab35-df28-46dd-95ad-725f8a812b79.png">

- [How to create memoized selectors with `reselect` to avoid heavy computations and creation of new refs at every store update, in order to avoid unnecessary re-renders of components](https://github.com/reduxjs/reselect#createselectorinputselectors--inputselectors-resultfunc-selectoroptions).

- [How React.memo works](https://reactjs.org/docs/react-api.html#reactmemo)
> <img width="686" alt="Screenshot 2023-02-03 at 10 30 09" src="https://user-images.githubusercontent.com/91890529/216564084-58b315a7-8a56-4235-a6cc-fde86e667cc0.png">


### ❓ Context

- **Impacted projects**: `mobile` `common` <!-- The list of end user projects impacted by the change. -->
- **Linked resource(s)**: [LIVE-4965] [LIVE-3986] <!-- Attach any ticket number if relevant. (JIRA / Github issue number) -->

### ✅ Checklist

- [ ] **Test coverage** <!-- Are your changes covered by tests? Features must be tested, bugfixes must include a test that would have detected the issue. -->
- [x] **Atomic delivery** <!-- Is this pull request standalone? In order words, does it depend on nothing else? Please explain if not checked. -->
- [x] **No breaking changes** <!-- If there are breaking changes, please explain why. -->

### 📸 Demo


Some results of a simple test where I used [react-native-flipper-performance-monitor](https://github.com/bamlab/react-native-flipper-performance-monitor) to get a score on a simple test scenario: opening an instance of the app with many accounts and measuring the FPS for 20 seconds:

<img width="751" alt="Screenshot 2023-02-03 at 18 13 25" src="https://user-images.githubusercontent.com/91890529/216665354-317f7d03-1050-4355-ae88-8321b184900f.png">



### 🚀 Expectations to reach

_Please make sure you follow these [**Important Steps**](https://github.com/LedgerHQ/ledger-live/blob/develop/CONTRIBUTING.md#important-steps)._

_Pull Requests must pass the CI and be internally validated in order to be merged._

<!-- If any of the expectations are not met please explain the reason in detail. -->


[LIVE-4964]: https://ledgerhq.atlassian.net/browse/LIVE-4964?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[LIVE-3986]: https://ledgerhq.atlassian.net/browse/LIVE-3986?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[LIVE-4965]: https://ledgerhq.atlassian.net/browse/LIVE-4965?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ