@mcayuelas-ledger 2023-02-07
LGTM on the part i used to work ! :) 

Nice job 🚀