@ofreyssinet-ledger 2023-02-13
@nabil-brn Thanks! I didn't know QA engineers were capable of manipulating "human feeling" 😄 :trollface: